[feat] Move 엔티티에 작성된 enum 경로들 엔티티와 패키지 동일화
- 현재 kapt 엔티티 필드로 enum 사용할때 enum의 패키지 위치가 엔티티와 다르면 Q클래스 생성시 필드 인식을 못함
  임시 방편으로 동일한 패키지 위치
- 추후 다른 방법 찾을때 다시 enum 패키지 분리
