+++ src/main/kotlin/io/runnershigh/backend/training/dto/response/ReadTrainingSchedule.kt
@@ -1,7 +1,7 @@
 package io.runnershigh.backend.training.dto.response
 
-import io.runnershigh.backend.training.entity.enum.TrainingColor
-import io.runnershigh.backend.training.entity.enum.TrainingStatus
+import io.runnershigh.backend.training.entity.TrainingColor
+import io.runnershigh.backend.training.entity.TrainingStatus
 import java.time.LocalDate
 
 data class ReadTrainingSchedule(
