+++ src/test/kotlin/io/runnershigh/backend/fixture/UserFixture.kt
@@ -1,9 +1,9 @@
 package io.runnershigh.backend.fixture
 
 import io.runnershigh.backend.user.entity.UserEntity
-import io.runnershigh.backend.user.entity.enum.AgeGroup
-import io.runnershigh.backend.user.entity.enum.Gender
-import io.runnershigh.backend.user.entity.enum.UserStatus
+import io.runnershigh.backend.user.entity.AgeGroup
+import io.runnershigh.backend.user.entity.Gender
+import io.runnershigh.backend.user.entity.UserStatus
 import net.datafaker.Faker
 import java.util.*
