+++ src/test/kotlin/io/runnershigh/backend/user/service/UserServiceImplTest.kt
@@ -14,8 +14,8 @@ import io.runnershigh.backend.shared.security.jwt.JwtTokenProvider
 import io.runnershigh.backend.user.dto.request.LoginRequest
 import io.runnershigh.backend.user.dto.request.SignupRequest
 import io.runnershigh.backend.user.entity.UserEntity
-import io.runnershigh.backend.user.entity.enum.AgeGroup
-import io.runnershigh.backend.user.entity.enum.Gender
+import io.runnershigh.backend.user.entity.AgeGroup
+import io.runnershigh.backend.user.entity.Gender
 import io.runnershigh.backend.user.exception.UserException
 import io.runnershigh.backend.user.exception.UserExceptionType
 import io.runnershigh.backend.user.repository.UserRepository
